ProtocolContext delegates to an instance of AbstractProtocolContextImplementation
location cache is an object
