+++ api/src/opentrons/protocol_api/module_contexts.py
@@ -1,11 +1,14 @@
 import asyncio
 import logging
-from typing import Generic, List, Optional, TYPE_CHECKING, TypeVar
+from typing import Generic, List, Optional, TYPE_CHECKING, TypeVar, Union
 
 from opentrons import types, commands as cmds
+from opentrons.broker import Broker
 from opentrons.hardware_control import modules
 from opentrons.hardware_control.types import Axis
 from opentrons.commands import CommandPublisher
+from opentrons.protocol_api.implementation.interfaces.protocol_context import \
+    AbstractProtocolContext
 from opentrons.protocols.types import APIVersion
 
 from .labware import (
@@ -16,8 +19,14 @@ from opentrons.protocols.geometry import planning
 from opentrons.protocols.api_support.util import requires_version
 
 if TYPE_CHECKING:
-    from .protocol_context import ProtocolContext
     from opentrons_shared_data.labware.dev_types import (LabwareDefinition)
+
+
+ModuleTypes = Union[
+    'TemperatureModuleContext',
+    'MagneticModuleContext',
+    'ThermocyclerContext'
+]
 
 ENGAGE_HEIGHT_UNIT_CNV = 2
 STANDARD_MAGDECK_LABWARE = [
@@ -37,7 +46,8 @@ class ModuleContext(CommandPublisher, Generic[GeometryType]):  # noqa(E302)
     """
 
     def __init__(self,
-                 ctx: 'ProtocolContext',
+                 ctx: AbstractProtocolContext,
+                 broker: Broker,
                  geometry: GeometryType,
                  at_version: APIVersion) -> None:
         """ Build the ModuleContext.
@@ -48,7 +58,7 @@ class ModuleContext(CommandPublisher, Generic[GeometryType]):  # noqa(E302)
         :param ctx: The parent context for the module
         :param geometry: The :py:class:`.ModuleGeometry` for the module
         """
-        super().__init__(ctx.broker)
+        super().__init__(broker)
         self._geometry = geometry
         self._ctx = ctx
         self._api_version = at_version
@@ -70,7 +80,7 @@ class ModuleContext(CommandPublisher, Generic[GeometryType]):  # noqa(E302)
         :returns: The properly-linked labware object
         """
         mod_labware = self._geometry.add_labware(labware)
-        self._ctx.deck.recalculate_high_z()
+        self._ctx.get_deck().recalculate_high_z()
         return mod_labware
 
     @requires_version(2, 0)
@@ -104,8 +114,8 @@ class ModuleContext(CommandPublisher, Generic[GeometryType]):  # noqa(E302)
                 'are trying to utilize new load_labware parameters in 2.1')
         lw = load(name, self._geometry.location,
                   label, namespace, version,
-                  bundled_defs=self._ctx._bundled_labware,
-                  extra_defs=self._ctx._extra_labware)
+                  bundled_defs=self._ctx.get_bundled_labware(),
+                  extra_defs=self._ctx.get_extra_labware())
         return self.load_labware_object(lw)
 
     @requires_version(2, 0)
@@ -191,14 +201,16 @@ class TemperatureModuleContext(ModuleContext[ModuleGeometry]):
     .. versionadded:: 2.0
 
     """
-    def __init__(self, ctx: 'ProtocolContext',
+    def __init__(self,
+                 ctx: AbstractProtocolContext,
+                 broker: Broker,
                  hw_module: modules.tempdeck.TempDeck,
                  geometry: ModuleGeometry,
                  at_version: APIVersion,
                  loop: asyncio.AbstractEventLoop) -> None:
         self._module = hw_module
         self._loop = loop
-        super().__init__(ctx, geometry, at_version)
+        super().__init__(ctx, broker, geometry, at_version)
 
     @cmds.publish.both(command=cmds.tempdeck_set_temp)
     @requires_version(2, 0)
@@ -273,14 +285,15 @@ class MagneticModuleContext(ModuleContext[ModuleGeometry]):
 
     """
     def __init__(self,
-                 ctx: 'ProtocolContext',
+                 ctx: AbstractProtocolContext,
+                 broker: Broker,
                  hw_module: modules.magdeck.MagDeck,
                  geometry: ModuleGeometry,
                  at_version: APIVersion,
                  loop: asyncio.AbstractEventLoop) -> None:
         self._module = hw_module
         self._loop = loop
-        super().__init__(ctx, geometry, at_version)
+        super().__init__(ctx, broker, geometry, at_version)
 
     @cmds.publish.both(command=cmds.magdeck_calibrate)
     @requires_version(2, 0)
@@ -344,7 +357,7 @@ class MagneticModuleContext(ModuleContext[ModuleGeometry]):
         if height is not None:
             dist = height
         elif height_from_base is not None and\
-                self._ctx._api_version >= APIVersion(2, 2):
+                self._ctx.get_api_version() >= APIVersion(2, 2):
             dist = height_from_base +\
                 modules.magdeck.OFFSET_TO_LABWARE_BOTTOM[self._module.model()]
         elif self.labware and self.labware.magdeck_engage_height is not None:
@@ -371,7 +384,7 @@ class MagneticModuleContext(ModuleContext[ModuleGeometry]):
 
         engage_height = self.labware.magdeck_engage_height
 
-        is_api_breakpoint = (self._ctx._api_version >= APIVersion(2, 3))
+        is_api_breakpoint = (self._ctx.get_api_version() >= APIVersion(2, 3))
         is_v1_module = (self._module.model() == 'magneticModuleV1')
         is_standard_lw = self.labware.load_name in STANDARD_MAGDECK_LABWARE
 
@@ -405,18 +418,19 @@ class ThermocyclerContext(ModuleContext[ThermocyclerGeometry]):
     .. versionadded:: 2.0
     """
     def __init__(self,
-                 ctx: 'ProtocolContext',
+                 ctx: AbstractProtocolContext,
+                 broker: Broker,
                  hw_module: modules.thermocycler.Thermocycler,
                  geometry: ThermocyclerGeometry,
                  at_version: APIVersion,
                  loop: asyncio.AbstractEventLoop) -> None:
         self._module = hw_module
         self._loop = loop
-        super().__init__(ctx, geometry, at_version)
+        super().__init__(ctx, broker, geometry, at_version)
 
     def _prepare_for_lid_move(self):
         loaded_instruments = [instr for mount, instr in
-                              self._ctx.loaded_instruments.items()
+                              self._ctx.get_loaded_instruments().items()
                               if instr is not None]
         try:
             instr = loaded_instruments[0]
@@ -425,10 +439,10 @@ class ThermocyclerContext(ModuleContext[ThermocyclerGeometry]):
                 "Cannot assure a safe gantry position to avoid colliding"
                 " with the lid of the Thermocycler Module.")
         else:
-            self._ctx._hw_manager.hardware.retract(instr._mount)
-            high_point = self._ctx._hw_manager.hardware.current_position(
+            self._ctx.get_hardware().hardware.retract(instr._mount)
+            high_point = self._ctx.get_hardware().hardware.current_position(
                     instr._mount)
-            trash_top = self._ctx.fixed_trash.wells()[0].top()
+            trash_top = self._ctx.get_fixed_trash().wells()[0].top()
             safe_point = trash_top.point._replace(
                     z=high_point[Axis.by_mount(instr._mount)])
             instr.move_to(types.Location(safe_point, None), force_direct=True)
