+++ api/src/opentrons/protocol_api/protocol_context.py
@@ -1,45 +1,27 @@
-import asyncio
 import contextlib
 import logging
-from typing import (Dict, Iterator, List,
-                    Optional, Set, Tuple, Union, TYPE_CHECKING)
+from typing import (Dict, List, Optional, Union, TYPE_CHECKING)
 
-from opentrons import types, commands as cmds
-from opentrons.hardware_control import (SynchronousAdapter, modules,
-                                        API, ExecutionManager)
-from opentrons.config import feature_flags as fflags
-from opentrons.commands import CommandPublisher
-from opentrons.protocols.api_support.constants import SHORT_TRASH_DECK, \
-    STANDARD_DECK
-from opentrons.protocols.types import APIVersion, Protocol
-from .labware import (
-    Labware, get_labware_definition, load_from_definition)
-from opentrons.protocols.geometry.module_geometry import (
-    ModuleGeometry, load_module, resolve_module_model,
-    resolve_module_type, models_compatible, ModuleType,
-    module_model_from_string)
+from opentrons import types
+from opentrons.hardware_control import API
+from opentrons.protocol_api.implementation.interfaces.protocol_context import \
+    AbstractProtocolContext
+from opentrons.protocols.types import APIVersion
+from .labware import Labware
+from opentrons.protocols.geometry.module_geometry import ModuleGeometry
 from opentrons.protocols.geometry.deck import Deck
-from opentrons.protocols.api_support.definitions import MAX_SUPPORTED_VERSION
 from .instrument_context import InstrumentContext
-from .module_contexts import (
-    ModuleContext, MagneticModuleContext, TemperatureModuleContext,
-    ThermocyclerContext)
+from .module_contexts import ModuleContext, ModuleTypes
 from opentrons.protocols.api_support.util import (
-    AxisMaxSpeeds, HardwareManager, requires_version, HardwareToManage,
-    APIVersionError, convert_door_state_to_bool)
+    AxisMaxSpeeds, requires_version)
+
 if TYPE_CHECKING:
     from opentrons_shared_data.labware.dev_types import LabwareDefinition
 
 MODULE_LOG = logging.getLogger(__name__)
 
-ModuleTypes = Union[
-    'TemperatureModuleContext',
-    'MagneticModuleContext',
-    'ThermocyclerContext'
-]
 
-
-class ProtocolContext(CommandPublisher):
+class ProtocolContext:
     """ The Context class is a container for the state of a protocol.
 
     It encapsulates many of the methods formerly found in the Robot class,
@@ -56,85 +38,12 @@ class ProtocolContext(CommandPublisher):
 
     """
 
-    def __init__(self,
-                 loop: asyncio.AbstractEventLoop = None,
-                 hardware: HardwareToManage = None,
-                 broker=None,
-                 bundled_labware: Dict[str, 'LabwareDefinition'] = None,
-                 bundled_data: Dict[str, bytes] = None,
-                 extra_labware: Dict[str, 'LabwareDefinition'] = None,
-                 api_version: APIVersion = None,
-                 ) -> None:
+    def __init__(self, implementation: AbstractProtocolContext) -> None:
         """ Build a :py:class:`.ProtocolContext`.
 
-        :param loop: An event loop to use. If not specified, this ctor will
-                     (eventually) call :py:meth:`asyncio.get_event_loop`.
-        :param hardware: An optional hardware controller to link to. If not
-                         specified, a new simulator will be created.
-        :param broker: An optional command broker to link to. If not
-                      specified, a dummy one is used.
-        :param bundled_labware: A dict mapping labware URIs to definitions.
-                                This is used when executing bundled protocols,
-                                and if specified will be the only allowed
-                                source for labware definitions, excluding the
-                                built in definitions and anything in
-                                ``extra_labware``.
-        :param bundled_data: A dict mapping filenames to the contents of data
-                             files. Can be used by the protocol, since it is
-                             exposed as
-                             :py:attr:`.ProtocolContext.bundled_data`
-        :param extra_labware: A dict mapping labware URIs to definitions. These
-                              URIs are searched during :py:meth:`.load_labware`
-                              in addition to the system definitions (if
-                              ``bundled_labware`` was not specified). Used to
-                              provide custom labware definitions.
-        :param api_version: The API version to use. If this is ``None``, uses
-                            the max supported version.
+        :param implementation: The implementation of the protocol API context
         """
-        super().__init__(broker)
-
-        self._api_version = api_version or MAX_SUPPORTED_VERSION
-        if self._api_version > MAX_SUPPORTED_VERSION:
-            raise RuntimeError(
-                f'API version {self._api_version} is not supported by this '
-                f'robot software. Please either reduce your requested API '
-                f'version or update your robot.')
-        self._loop = loop or asyncio.get_event_loop()
-        deck_load_name = SHORT_TRASH_DECK if fflags.short_fixed_trash() \
-            else STANDARD_DECK
-        self._deck_layout = Deck(load_name=deck_load_name)
-        self._instruments: Dict[types.Mount, Optional[InstrumentContext]]\
-            = {mount: None for mount in types.Mount}
-        self._modules: Set[ModuleContext] = set()
-        self._last_moved_instrument: Optional[types.Mount] = None
-        self._location_cache: Optional[types.Location] = None
-
-        self._hw_manager = HardwareManager(hardware)
-        self._log = MODULE_LOG.getChild(self.__class__.__name__)
-        self._commands: List[str] = []
-        self._unsubscribe_commands = None
-        self.clear_commands()
-
-        self._bundled_labware = bundled_labware
-        self._extra_labware = extra_labware or {}
-
-        self._bundled_data: Dict[str, bytes] = bundled_data or {}
-        self._default_max_speeds = AxisMaxSpeeds()
-
-    @classmethod
-    def build_using(cls,
-                    protocol: Protocol,
-                    *args, **kwargs):
-        """ Build an API instance for the specified parsed protocol
-
-        This is used internally to provision the context with bundle
-        contents or api levels.
-        """
-        kwargs['bundled_data'] = getattr(protocol, 'bundled_data', None)
-        kwargs['bundled_labware'] = getattr(protocol, 'bundled_labware', None)
-        kwargs['api_version'] = getattr(
-            protocol, 'api_level', MAX_SUPPORTED_VERSION)
-        return cls(*args, **kwargs)
+        self._implementation = implementation
 
     @property  # type: ignore
     @requires_version(2, 0)
@@ -146,7 +55,7 @@ class ProtocolContext(CommandPublisher):
         by the robot software. For the highest version supported by the
         robot software, see :py:attr:`.protocol_api.MAX_SUPPORTED_VERSION`.
         """
-        return self._api_version
+        return self._implementation.get_api_version()
 
     @property  # type: ignore
     @requires_version(2, 0)
@@ -158,17 +67,11 @@ class ProtocolContext(CommandPublisher):
         ``data/mydata/aspirations.csv`` it will be in the dict as
         ``'aspirations.csv'``) to the bytes contents of the files.
         """
-        return self._bundled_data
+        return self._implementation.get_bundled_data()
 
     def cleanup(self):
         """ Finalize and clean up the protocol context. """
-        if self._unsubscribe_commands:
-            self._unsubscribe_commands()
-            self._unsubscribe_commands = None
-
-    def __del__(self):
-        if getattr(self, '_unsubscribe_commands', None):
-            self._unsubscribe_commands()  # type: ignore
+        self._implementation.cleanup()
 
     @property  # type: ignore
     @requires_version(2, 0)
@@ -201,29 +104,15 @@ class ProtocolContext(CommandPublisher):
                 protocol.max_speeds['X'] = None  # reset to default
 
         """
-        return self._default_max_speeds
+        return self._implementation.get_max_speeds()
 
     @requires_version(2, 0)
     def commands(self):
-        return self._commands
+        return self._implementation.get_commands()
 
     @requires_version(2, 0)
     def clear_commands(self):
-        self._commands.clear()
-        if self._unsubscribe_commands:
-            self._unsubscribe_commands()
-
-        def on_command(message):
-            payload = message.get('payload')
-            text = payload.get('text')
-            if text is None:
-                return
-
-            if message['$'] == 'before':
-                self._commands.append(text.format(**payload))
-
-        self._unsubscribe_commands = self.broker.subscribe(
-            cmds.types.COMMAND, on_command)
+        self._implementation.clear_commands()
 
     @contextlib.contextmanager
     def temp_connect(self, hardware: API):
@@ -241,25 +130,7 @@ class ProtocolContext(CommandPublisher):
             # an error occurred in the code inside the with block
 
         """
-        old_hw = self._hw_manager.hardware
-        old_tc = None
-        tc_context = None
-        try:
-            self._hw_manager.set_hw(hardware)
-            for mod_ctx in self._modules:
-                if isinstance(mod_ctx, ThermocyclerContext):
-                    tc_context = mod_ctx
-                    hw_tc = next(hw_mod for hw_mod in
-                                 hardware.attached_modules
-                                 if hw_mod.name() == 'thermocycler')
-                    if hw_tc:
-                        old_tc = mod_ctx._module
-                        mod_ctx._module = hw_tc
-            yield self
-        finally:
-            self._hw_manager.set_hw(old_hw)
-            if tc_context is not None and old_tc is not None:
-                tc_context._module = old_tc
+        self._implementation.temp_connect(hardware)
 
     @requires_version(2, 0)
     def connect(self, hardware: API):
@@ -271,18 +142,17 @@ class ProtocolContext(CommandPublisher):
         :py:class:`.ProtocolContext`; :py:meth:`disconnect` simply creates
         a new simulator and replaces the current hardware with it.
         """
-        self._hw_manager.set_hw(hardware)
-        self._hw_manager.hardware.cache_instruments()
+        self._implementation.connect(hardware)
 
     @requires_version(2, 0)
     def disconnect(self):
         """ Disconnect from currently-connected hardware and simulate instead
         """
-        self._hw_manager.reset_hw()
+        self._implementation.disconnect()
 
     @requires_version(2, 0)
     def is_simulating(self) -> bool:
-        return self._hw_manager.hardware.is_simulator
+        return self._implementation.is_simulating()
 
     @requires_version(2, 0)
     def load_labware_from_definition(
@@ -305,10 +175,9 @@ class ProtocolContext(CommandPublisher):
                           as in the run log and the calibration view in the
                           Opentrons app.
         """
-        parent = self.deck.position_for(location)
-        labware_obj = load_from_definition(labware_def, parent, label)
-        self._deck_layout[location] = labware_obj
-        return labware_obj
+        return self._implementation.load_labware_from_definition(labware_def,
+                                                                 location,
+                                                                 label)
 
     @requires_version(2, 0)
     def load_labware(
@@ -341,12 +210,8 @@ class ProtocolContext(CommandPublisher):
         :param int version: The version of the labware definition. If
             unspecified, will use version 1.
         """
-        labware_def = get_labware_definition(
-            load_name, namespace, version,
-            bundled_defs=self._bundled_labware,
-            extra_defs=self._extra_labware)
-        return self.load_labware_from_definition(
-            labware_def, location, label)
+        return self._implementation.load_labware(load_name, location, label,
+                                                 namespace, version)
 
     @requires_version(2, 0)
     def load_labware_by_name(
@@ -383,16 +248,7 @@ class ProtocolContext(CommandPublisher):
         :returns: Dict mapping deck slot number to labware, sorted in order of
                   the locations.
         """
-        def _only_labwares() -> Iterator[
-                Tuple[int, Union[Labware, ModuleGeometry]]]:
-            for slotnum, slotitem in self._deck_layout.items():
-                if isinstance(slotitem, Labware):
-                    yield slotnum, slotitem
-                elif isinstance(slotitem, ModuleGeometry):
-                    if slotitem.labware:
-                        yield slotnum, slotitem.labware
-
-        return dict(_only_labwares())
+        return self._implementation.get_loaded_labwares()
 
     @requires_version(2, 0)
     def load_module(
@@ -427,58 +283,9 @@ class ProtocolContext(CommandPublisher):
         :returns ModuleContext: The loaded and initialized
                                 :py:class:`ModuleContext`.
         """
-        resolved_model = resolve_module_model(module_name)
-        resolved_type = resolve_module_type(resolved_model)
-        resolved_location = self._deck_layout.resolve_module_location(
-            resolved_type, location)
-        if self._api_version < APIVersion(2, 4) and configuration:
-            raise APIVersionError(
-                f'You have specified API {self._api_version}, but you are'
-                'using thermocycler parameters only available in 2.4')
-
-        geometry = load_module(
-            resolved_model,
-            self._deck_layout.position_for(
-                resolved_location),
-            self._api_version, configuration)
-
-        hc_mod_instance = None
-        mod_class = {
-            ModuleType.MAGNETIC: MagneticModuleContext,
-            ModuleType.TEMPERATURE: TemperatureModuleContext,
-            ModuleType.THERMOCYCLER: ThermocyclerContext}[resolved_type]
-        for mod in self._hw_manager.hardware.attached_modules:
-            if models_compatible(
-                    module_model_from_string(mod.model()), resolved_model):
-                hc_mod_instance = SynchronousAdapter(mod)
-                break
-
-        if self.is_simulating() and hc_mod_instance is None:
-            mod_type = {
-                ModuleType.MAGNETIC: modules.magdeck.MagDeck,
-                ModuleType.TEMPERATURE: modules.tempdeck.TempDeck,
-                ModuleType.THERMOCYCLER: modules.thermocycler.Thermocycler
-                }[resolved_type]
-            hc_mod_instance = SynchronousAdapter(mod_type(
-                    port='',
-                    simulating=True,
-                    loop=self._hw_manager.hardware.loop,
-                    execution_manager=ExecutionManager(
-                        loop=self._hw_manager.hardware.loop),
-                    sim_model=resolved_model.value))
-            hc_mod_instance._connect()
-        if hc_mod_instance:
-            mod_ctx = mod_class(self,
-                                hc_mod_instance,
-                                geometry,
-                                self.api_version,
-                                self._loop)
-        else:
-            raise RuntimeError(
-                f'Could not find specified module: {module_name}')
-        self._modules.add(mod_ctx)
-        self._deck_layout[resolved_location] = geometry
-        return mod_ctx
+        return self._implementation.load_module(module_name=module_name,
+                                                location=location,
+                                                configuration=configuration)
 
     @property  # type: ignore
     @requires_version(2, 0)
@@ -496,11 +303,7 @@ class ProtocolContext(CommandPublisher):
                                            contexts. The elements may not be
                                            ordered by slot number.
         """
-        def _modules() -> Iterator[Tuple[int, 'ModuleContext']]:
-            for module in self._modules:
-                yield int(module.geometry.parent), module
-
-        return dict(_modules())
+        return self._implementation.get_loaded_modules()
 
     @requires_version(2, 0)
     def load_instrument(
@@ -546,32 +349,11 @@ class ProtocolContext(CommandPublisher):
             raise TypeError(
                 "mount should be either an instance of opentrons.types.Mount"
                 " or a string, but is {}.".format(mount))
-        self._log.info("Trying to load {} on {} mount"
-                       .format(instrument_name, checked_mount.name.lower()))
-        instr = self._instruments[checked_mount]
-        if instr and not replace:
-            raise RuntimeError("Instrument already present in {} mount: {}"
-                               .format(checked_mount.name.lower(),
-                                       instr.name))
-        attached = {att_mount: instr.get('model', None)
-                    for att_mount, instr
-                    in self._hw_manager.hardware.attached_instruments.items()}
-        attached[checked_mount] = instrument_name
-        self._log.debug("cache instruments expectation: {}"
-                        .format(attached))
-        self._hw_manager.hardware.cache_instruments(attached)
-        # If the cache call didn’t raise, the instrument is attached
-        new_instr = InstrumentContext(
-            ctx=self,
-            hardware_mgr=self._hw_manager,
-            mount=checked_mount,
-            at_version=self._api_version,
-            tip_racks=tip_racks,
-            log_parent=self._log,
-            requested_as=instrument_name)
-        self._instruments[checked_mount] = new_instr
-        self._log.info("Instrument {} loaded".format(new_instr))
-        return new_instr
+
+        return self._implementation.load_instrument(instrument_name=instrument_name,
+                                                    mount=checked_mount,
+                                                    tip_racks=tip_racks,
+                                                    replace=replace)
 
     @property  # type: ignore
     @requires_version(2, 0)
@@ -589,11 +371,8 @@ class ProtocolContext(CommandPublisher):
                   in that mount. If no instrument is loaded in the mount,
                   it will not be present
         """
-        return {mount.name.lower(): instr for mount, instr
-                in self._instruments.items()
-                if instr}
+        return self._implementation.get_loaded_instruments()
 
-    @cmds.publish.both(command=cmds.pause)
     @requires_version(2, 0)
     def pause(self, msg=None):
         """ Pause execution of the protocol until resume is called.
@@ -604,15 +383,13 @@ class ProtocolContext(CommandPublisher):
 
         :param str msg: A message to echo back to connected clients.
         """
-        self._hw_manager.hardware.pause()
+        self._implementation.pause(msg=msg)
 
-    @cmds.publish.both(command=cmds.resume)
     @requires_version(2, 0)
     def resume(self):
         """ Resume a previously-paused protocol """
-        self._hw_manager.hardware.resume()
+        self._implementation.resume()
 
-    @cmds.publish.both(command=cmds.comment)
     @requires_version(2, 0)
     def comment(self, msg):
         """
@@ -623,9 +400,8 @@ class ProtocolContext(CommandPublisher):
         so cannot be used to communicate real-time information from the robot's
         actual run.
         """
-        pass
+        self._implementation.comment(msg=msg)
 
-    @cmds.publish.both(command=cmds.delay)
     @requires_version(2, 0)
     def delay(self, seconds=0, minutes=0, msg=None):
         """ Delay protocol execution for a specific amount of time.
@@ -636,25 +412,13 @@ class ProtocolContext(CommandPublisher):
         If both `seconds` and `minutes` are specified, they will be added.
         """
         delay_time = seconds + minutes * 60
-        self._hw_manager.hardware.delay(delay_time)
+        self._implementation.delay(seconds=delay_time, msg=msg)
 
     @requires_version(2, 0)
     def home(self):
         """ Homes the robot.
         """
-        self._log.debug("home")
-        self._location_cache = None
-        self._hw_manager.hardware.home()
-
-    @property
-    def location_cache(self) -> Optional[types.Location]:
-        """ The cache used by the robot to determine where it last was.
-        """
-        return self._location_cache
-
-    @location_cache.setter
-    def location_cache(self, loc: Optional[types.Location]):
-        self._location_cache = loc
+        self._implementation.home()
 
     @property  # type: ignore
     @requires_version(2, 0)
@@ -678,7 +442,7 @@ class ProtocolContext(CommandPublisher):
         reflect the labware physically on the deck (or be higher than the
         labware on the deck).
         """
-        return self._deck_layout
+        return self._implementation.get_deck()
 
     @property  # type: ignore
     @requires_version(2, 0)
@@ -688,10 +452,7 @@ class ProtocolContext(CommandPublisher):
         It has one well and should be accessed like labware in your protocol.
         e.g. ``protocol.fixed_trash['A1']``
         """
-        trash = self._deck_layout['12']
-        if not trash:
-            raise RuntimeError("Robot must have a trash container in 12")
-        return trash  # type: ignore
+        return self._implementation.get_fixed_trash()
 
     @requires_version(2, 5)
     def set_rail_lights(self, on: bool):
@@ -700,17 +461,16 @@ class ProtocolContext(CommandPublisher):
 
         :param bool on: If true, turn on rail lights; otherwise, turn off.
         """
-        self._hw_manager.hardware.set_lights(rails=on)
+        self._implementation.set_rail_lights(on=on)
 
     @property  # type: ignore
     @requires_version(2, 5)
     def rail_lights_on(self) -> bool:
         """ Returns True if the rail lights are on """
-        return self._hw_manager.hardware.get_lights()['rails']
+        return self._implementation.get_rail_lights_on()
 
     @property  # type: ignore
     @requires_version(2, 5)
     def door_closed(self) -> bool:
         """ Returns True if the robot door is closed """
-        return convert_door_state_to_bool(
-            self._hw_manager.hardware.door_state)
+        return self._implementation.door_closed()
